BEAM-8017: Added error to urnWindowCoder
Currently, the urnToWindowCoder function was set to panic when the
default case is reached. i.e. an unexpected URN is received.

In this patch, we change the return signature of the function to
return an error. This function is only used by the WindowCoder
function; and we also patch the appropriate handling of the error
in that function as well.